static inner class 였던 CafeDetails가 다른 API에서도 재사용 됨에 따라서 그냥 class로 전환
