+++ src/main/java/com/cafehub/backend/domain/cafe/dto/response/CafeListResponseDTO.java
@@ -1,5 +1,6 @@
 package com.cafehub.backend.domain.cafe.dto.response;
 
+import com.cafehub.backend.domain.cafe.dto.CafeDetails;
 import com.cafehub.backend.domain.cafe.entity.Theme;
 import com.querydsl.core.annotations.QueryProjection;
 import lombok.*;
@@ -11,37 +12,6 @@ import java.util.List;
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @Builder
 public class CafeListResponseDTO {
-
-    @Getter
-    @Setter
-    public static class CafeDetails {
-
-        private Long cafeId;
-
-        private String cafePhotoUrl;
-
-        private String cafeName;
-
-        private Theme cafeTheme;
-
-        private Double cafeRating;
-
-        private Integer cafeReviewNum;
-
-
-        @QueryProjection
-        @Builder
-        public CafeDetails(Long cafeId, String cafePhotoUrl, String cafeName, Theme cafeTheme, Double cafeRating, Integer cafeReviewNum) {
-            this.cafeId = cafeId;
-            this.cafePhotoUrl = cafePhotoUrl;
-            this.cafeName = cafeName;
-            this.cafeTheme = cafeTheme;
-            this.cafeRating = cafeRating;
-            this.cafeReviewNum = cafeReviewNum;
-        }
-    }
-
-
 
     private List<CafeDetails> cafeList;
