+++ src/main/java/com/cafehub/backend/domain/cafe/repository/CafeRepositoryCustom.java
@@ -1,5 +1,6 @@
 package com.cafehub.backend.domain.cafe.repository;
 
+import com.cafehub.backend.domain.cafe.dto.CafeDetails;
 import com.cafehub.backend.domain.cafe.dto.request.CafeListRequestDTO;
 import com.cafehub.backend.domain.cafe.dto.response.CafeInfoResponseDTO;
 import com.cafehub.backend.domain.cafe.dto.response.CafeListResponseDTO;
@@ -9,6 +10,6 @@ import java.util.List;
 
 public interface CafeRepositoryCustom {
 
-    Slice<CafeListResponseDTO.CafeDetails> findCafesBySlice (CafeListRequestDTO requestDTO);
+    Slice<CafeDetails> findCafesBySlice (CafeListRequestDTO requestDTO);
 
 }
